apigateway/iot_telemetry_rest/requirements.txt
schema==0.6.8
psycopg2-binary==2.7.6
